new project/requirements.txt
flask==3.0.3
requests==2.32.3
docker==7.1.0
paramiko==3.4.0
numpy==1.26.4
scikit-learn==1.5.1
reportlab==4.2.0
cryptography==42.0.8
opa-client==0.7.1
hvac==2.3.0
websocket-client==1.8.0
pydantic==2.8.2
gitpython==3.1.43
pytest==8.3.2
prometheus-client==0.20.0
black==24.4.2
ruff==0.5.5
mypy==1.10.0
PyJWT==2.9.0
fastapi==0.112.2
uvicorn==0.30.5
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
pydantic-settings==2.6.1
